+++ crates/editor/src/display_map.rs
@@ -183,6 +183,7 @@ pub enum HighlightKey {
     PickerPreview,
     ProjectSearchView,
     Rename,
+    SearchPanelMatches,
     SearchWithinRange,
     SelectedTextHighlight,
     SyntaxTreeView(usize),